[ADF-NO-ISSUE] Removing wrong dependencies
