+++ lib/testing/src/lib/core/material/public-api.ts
@@ -15,21 +15,4 @@
  * limitations under the License.
  */
 
-import { element, by } from 'protractor';
-import { BrowserVisibility } from '../core/browser-visibility';
-
-export class TabsPage {
-
-    clickTabByTitle(tabTitle) {
-        let tab = element(by.cssContainingText("div[id*='mat-tab-label']", tabTitle));
-        BrowserVisibility.waitUntilElementIsVisible(tab);
-        tab.click();
-    }
-
-    checkTabIsSelectedByTitle(tabTitle) {
-        let tab = element(by.cssContainingText("div[id*='mat-tab-label']", tabTitle));
-        tab.getAttribute('aria-selected').then((result) => {
-            expect(result).toBe('true');
-        });
-    }
-}
+export * from './tabs.page';
