--- lib/testing/src/lib/material/public-api.ts
@@ -1,5 +1,0 @@
-/*
- * Public API Surface of testing
- */
-
-export * from './tabs.page';
