+++ lib/testing/src/lib/process-services-cloud/pages/public-api.ts
@@ -1,9 +1,20 @@
-/*
- * Public API Surface of testing
+/*!
+ * @license
+ * Copyright 2019 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 export * from './login-sso.page';
-export * from '../../core/pages/user-info.page';
-export * from '../app/app-list-cloud.page';
 export * from './start-tasks-cloud-component.page';
 export * from './task-header-cloud-component.page';
