+++ lib/testing/src/public-api.ts
@@ -1,9 +1,22 @@
-/*
- * Public API Surface of testing
+/*!
+ * @license
+ * Copyright 2019 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 export * from './lib/core/public-api';
-export * from './lib/material/public-api';
+export * from './lib/core/material/public-api';
 export * from './lib/content-services/public-api';
 export * from './lib/process-services/public-api';
 export * from './lib/process-services-cloud/public-api';
